requirements.txt
cryptography.fernet